requirements.txt
loguru==0.7.2
pandas==1.5.3
missingno==0.5.2
numpy==1.26.4
ppscore==1.3.0
matplotlib
ydata-profiling==4.9.0
scikit-learn==1.4.2
yellowbrick==1.5
pycaret==3.3.2
xgboost==1.6.2
pendulum==3.0.0
mlflow==2.15.0
papermill==2.6.0
pytest==8.3.2
jupyter
python-dotenv==1.0.1
numba==0.58.1
azure-ai-ml==1.18.0
azureml-mlflow
azure-identity
logtail-python
ipykernel
black
isort